+++ pac4j-core/src/main/java/org/pac4j/core/engine/CallbackLogic.java
@@ -1,10 +1,6 @@
 package org.pac4j.core.engine;
 
 import org.pac4j.core.config.Config;
-import org.pac4j.core.context.WebContext;
-import org.pac4j.core.context.session.SessionStore;
-import org.pac4j.core.http.adapter.HttpActionAdapter;
-import org.pac4j.core.profile.factory.ProfileManagerFactory;
 
 /**
  * Callback logic to finish the login process for an indirect client.
@@ -18,16 +14,12 @@ public interface CallbackLogic {
     /**
      * Perform the callback logic.
      *
-     * @param webContext the web context
-     * @param sessionStore the session store
-     * @param profileManagerFactory the profile manager factory
      * @param config the security configuration
-     * @param httpActionAdapter the HTTP action adapter
      * @param defaultUrl the default url
      * @param renewSession whether the session must be renewed
      * @param defaultClient the default client
+     * @param parameters additional parameters
      * @return the resulting action of the callback
      */
-    Object perform(WebContext webContext, SessionStore sessionStore, ProfileManagerFactory profileManagerFactory, Config config,
-                   HttpActionAdapter httpActionAdapter, String defaultUrl, Boolean renewSession, String defaultClient);
+    Object perform(Config config, String defaultUrl, Boolean renewSession, String defaultClient, Object... parameters);
 }