+++ pac4j-core/src/main/java/org/pac4j/core/util/security/SecurityEndpoint.java
@@ -1,8 +1,6 @@
 package org.pac4j.core.util.security;
 
 import org.pac4j.core.config.Config;
-import org.pac4j.core.engine.SecurityLogic;
-import org.pac4j.core.http.adapter.HttpActionAdapter;
 
 /**
  * Security endpoint.
@@ -17,10 +15,6 @@ public interface SecurityEndpoint {
     void setAuthorizers(String authorizers);
 
     void setMatchers(String matchers);
-
-    void setSecurityLogic(SecurityLogic securityLogic);
-
-    void setHttpActionAdapter(HttpActionAdapter httpActionAdapter);
 
     void setConfig(Config config);
 }